refactored WindowsAppDirs methods rtypes to path
