+++ src/main/kotlin/io/github/erayerdin/kappdirs/appdirs/WindowsAppDirs.kt
@@ -1,43 +1,35 @@
 package io.github.erayerdin.kappdirs.appdirs
 
-import io.github.erayerdin.kappdirs.AppDirs
 import java.nio.file.Path
 import java.nio.file.Paths
 
 internal class WindowsAppDirs: AppDirs {
-    override fun getUserDataDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): String {
-        val appData: String
-        val path: Path
-
-        if (roaming)
-            appData = APPDATA
+    override fun getUserDataDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): Path {
+        val appData: String = if (roaming)
+            APPDATA
         else
-            appData = LOCALAPPDATA
+            LOCALAPPDATA
 
-        path = Paths.get(appData, appAuthor, appName, appVersion)
-        return path.toString()
+        return Paths.get(appData, appAuthor, appName, appVersion)
     }
 
-    override fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): String {
+    override fun getUserConfigDir(appName: String, appVersion: String, appAuthor: String, roaming: Boolean): Path {
         return getUserDataDir(appName, appVersion, appAuthor, roaming)
     }
 
-    override fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String): String {
-        val path = Paths.get(LOCALAPPDATA, appAuthor, appName, "Cache", appVersion)
-        return path.toString()
+    override fun getUserCacheDir(appName: String, appVersion: String, appAuthor: String): Path {
+        return Paths.get(LOCALAPPDATA, appAuthor, appName, "Cache", appVersion)
     }
 
-    override fun getUserLogDir(appName: String, appVersion: String, appAuthor: String): String {
-        val path = Paths.get(LOCALAPPDATA, appAuthor, appName, "Logs", appVersion)
-        return path.toString()
+    override fun getUserLogDir(appName: String, appVersion: String, appAuthor: String): Path {
+        return Paths.get(LOCALAPPDATA, appAuthor, appName, "Logs", appVersion)
     }
 
-    override fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): String {
-        val path = Paths.get(PROGRAMDATA, appAuthor, appName, appVersion)
-        return path.toString()
+    override fun getSiteDataDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): Path {
+        return Paths.get(PROGRAMDATA, appAuthor, appName, appVersion)
     }
 
-    override fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): String {
+    override fun getSiteConfigDir(appName: String, appVersion: String, appAuthor: String, local: Boolean): Path {
         return getSiteDataDir(appName, appVersion, appAuthor, local)
     }
 }
