+++ src/test/kotlin/io/github/erayerdin/kappdirs/appdirs/UnixAppDirsTest.kt
@@ -1,6 +1,5 @@
 package io.github.erayerdin.kappdirs.appdirs
 
-import io.github.erayerdin.kappdirs.AppDirs
 import org.junit.*
 import org.junit.Assert.assertEquals
