+++ src/test/kotlin/io/github/erayerdin/kappdirs/appdirs/WindowsAppDirsTest.kt
@@ -1,6 +1,5 @@
 package io.github.erayerdin.kappdirs.appdirs
 
-import io.github.erayerdin.kappdirs.AppDirs
 import org.junit.Assert.assertTrue
 import org.junit.Assume
 import org.junit.Before
